removing old board class
